requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
docker==4.0.2
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
PyYAML==5.1.1
requests==2.22.0
six==1.12.0
urllib3==1.25.3
websocket-client==0.56.0
Werkzeug==0.15.5
